Fix warning and integrate function to StructDoc
