+++ mithril-common/src/generate_doc/extract_clap_info.rs
@@ -1,5 +1,3 @@
-use std::collections::HashMap;
-
 use clap::{Command, Arg, builder::StyledStr};
 
 use super::{StructDoc, FieldDoc};
@@ -37,29 +35,7 @@ pub fn extract_parameters(cmd: &Command) -> StructDoc {
     }
 }
 
-pub fn merge_struct_doc(s1: &StructDoc, s2: &StructDoc) -> StructDoc {
-   
-    let mut data_map1 = s1.data.iter()
-        .map(|field_doc| (field_doc.parameter.clone(), field_doc.clone()))
-        .collect::<HashMap<_,_>>();
 
-    for field_doc in s2.data.iter() {
-        if !data_map1.contains_key(&field_doc.parameter) {
-            data_map1.insert(field_doc.parameter.clone(), field_doc.clone());
-        } else {
-           
-            let mut d = data_map1.get(&field_doc.parameter).unwrap().clone();
-            if d.default_value.is_none() {
-                d.default_value = field_doc.default_value.clone();
-                data_map1.insert(field_doc.parameter.clone(), d);
-            }
-        }
-    }
-    let result = StructDoc {
-        data: data_map1.values().cloned().collect(),
-    };
-    result
-}
 
 #[cfg(test)]
 mod tests {
@@ -115,45 +91,5 @@ mod tests {
         }
     }
 
-    #[test]
-    fn test_merge_struct_doc() {
-        let s1 = {
-            let mut s = StructDoc::default();
-            s.add_param("A", "Param first A", Some("default A".to_string()));
-            s.add_param("B", "Param first B", None);
-            s.add_param("C", "Param first C", Some("default C".to_string()));
-            s.add_param("D", "Param first D", None);
-            s
-        };
 
-        let s2 = {
-            let mut s = StructDoc::default();
-            s.add_param("A", "Param second A", None);
-            s.add_param("B", "Param second B", Some("default B".to_string()));
-            s.add_param("E", "Param second E", None);
-            s.add_param("F", "Param second F", Some("default F".to_string()));
-            s
-        };
-
-        let result = merge_struct_doc(&s1, &s2);
-
-        let data = result.data;
-        let data_map = data.into_iter().map(|fieldDoc| (fieldDoc.parameter.clone(), fieldDoc)).collect::<HashMap<_,_>>();
-
-        assert_eq!(6, data_map.iter().count());
-        assert_eq!("Param first A", data_map.get("A").unwrap().description);
-        assert_eq!("Param first B", data_map.get("B").unwrap().description);
-        assert_eq!("Param first C", data_map.get("C").unwrap().description);
-        assert_eq!("Param first D", data_map.get("D").unwrap().description);
-        assert_eq!("Param second E", data_map.get("E").unwrap().description);
-        assert_eq!("Param second F", data_map.get("F").unwrap().description);
-
-        assert_eq!(Some("default A".to_string()), data_map.get("A").unwrap().default_value);
-        assert_eq!(Some("default B".to_string()), data_map.get("B").unwrap().default_value);
-        assert_eq!(Some("default C".to_string()), data_map.get("C").unwrap().default_value);
-        assert_eq!(None, data_map.get("D").unwrap().default_value);
-        assert_eq!(None, data_map.get("E").unwrap().default_value);
-        assert_eq!(Some("default F".to_string()), data_map.get("F").unwrap().default_value);
-
-    }
 }
